requirements.txt
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
pydantic-ai[openai]>=1.39.0
python-dotenv>=1.2.0
websockets>=15.0.0
python-multipart>=0.0.21
pandas==2.1.4
openpyxl==3.1.2
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.1.0
httpx>=0.24.0
openai>=1.0.0
norfair[video]
numpy
Pillow
matplotlib==3.6.1
PyYAML==6.0
tqdm==4.64.1
seaborn==0.12.0
ipython==8.5.0
ultralytics
colorama==0.4.5
requests==2.28.1
